refactor: rename file to match its default export
