+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-01T10:43:17.772Z\n"
-"PO-Revision-Date: 2019-02-01T10:43:17.772Z\n"
+"POT-Creation-Date: 2019-02-01T10:56:51.780Z\n"
+"PO-Revision-Date: 2019-02-01T10:56:51.780Z\n"
 
 msgid "Cancel"
 msgstr ""
