+++ src/pages/reporting-rate-summary/ReportingRateSummary.js
@@ -29,7 +29,7 @@ import PeriodPickerComponent from '../../components/period-picker-with-period-ty
 import Report from '../../components/report/Report';
 
 /* utils */
-import { getDocsUrl } from '../../helpers/docs';
+import { getDocsUrl } from '../../helpers/getDocsUrl';
 import { LOADING, SUCCESS } from '../../helpers/feedbackSnackBarTypes';
 
 /* styles */
